Refactor getNoteTitles (#1894)
This PR refactors getNoteTiles to make the code a little cleaner, add types to
function, and removes lodash dependency.
